feat: implement v0.47.0 — SHACL truthfulness, GUC validators, cache SRFs, fuzz hardening
- sh:lessThanOrEquals: add ShapeConstraint enum variant, parser cases, and
  check_less_than_or_equals() checker in shape_based.rs (v0.47.0 S4-4)
- sh:closed, sh:uniqueLang, sh:pattern: already implemented; wire confirmed
  via dispatch_constraint() and four new pg_regress tests (shacl_closed.sql,
  shacl_unique_lang.sql, shacl_pattern.sql, shacl_lt_or_equals.sql)
- six GUC check_hook validators: federation_on_error, federation_on_partial,
  sparql_overflow_action, tracing_exporter, embedding_index_type,
  embedding_precision — reject invalid values at SET time (S5-1)
- preallocate_sid_ranges() wired in run_inference_seminaive() when
  datalog_parallel_workers > 1; dead_code warning eliminated (S1-2)
- sparql/translate/ module split complete: sqlgen.rs 753 LoC (S2-3)
  bgp.rs, filter.rs, graph.rs, group.rs, join.rs, left_join.rs,
  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, disate  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distincal  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, disai  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distinctt_  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distinct.il  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs den  union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rs, distincte:   union.rs, distinct.rs ext  union.rs, distinct.rs ext  union.rsts/owl2rl/known_failures.txt updated (S3-3)
- promotion-race stress test: tests/stress/promotion_race.sh (S8-5)
- migration script: sql/- migration script: sql/- migration script: sql/- migration scriptar- migration script: sql/- migration script: sql/- migration scri ticked
- CHANGELOG.md: v0.47.0 entry added
