+++ pg_ripple.control
@@ -2,7 +2,7 @@
 # See: https://www.postgresql.org/docs/current/extend-extensions.html
 
 comment = 'High-performance RDF triple store with native SPARQL query execution'
-default_version = '0.46.0'
+default_version = '0.47.0'
 # MUST match the current release version (keep in sync with Cargo.toml)
 module_pathname = '$libdir/pg_ripple'
 relocatable = false
