+++ tests/owl2rl/known_failures.txt
@@ -1,12 +1,15 @@
-# OWL 2 RL known failures for pg_ripple (v0.46.0)
+# OWL 2 RL known failures for pg_ripple (v0.47.0)
 #
 # Format: one entry per line, prefixed with "owl2rl:" followed by the test ID.
 # Lines starting with "#" are comments.
 #
-# Tests listed here are skipped by the owl2rl_suite runner.  When a test is
-# fixed, remove its line from this file and re-run the suite to confirm it passes.
+# Tests listed here are treated as XFAIL in CI and do NOT block releases.
+# Removing a line (fixing the underlying issue) is cause for celebration.
 #
-# Example:
-#   owl2rl:rl-syntax-AnnotationProperty-001
+# v0.47.0 baseline: 62/66 pass (93.9%)
 #
-# Currently no known failures at initial release — file is a placeholder.
+# id                     added      target     reason
+owl2rl:prp-spo2         # v0.47.0  v0.49.0    three-hop propertyChainAxiom not yet supported
+owl2rl:scm-sco          # v0.47.0  v0.49.0    cyclical subclass → owl:equivalentClass not inferred
+owl2rl:eq-diff1         # v0.47.0  v0.50.0    owl:Nothing not propagated from sameAs/differentFrom conflict
+owl2rl:dt-type2         # v0.47.0  v0.51.0    XSD numeric promotion not applied on store
